// student-growth-system-design-v3/student-growth-system-design-v3/frontend/src/api/generated/models/ExamCreate.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Student Growth Archive API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 3.15.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { ExamType } from './ExamType';
import {
    ExamTypeFromJSON,
    ExamTypeFromJSONTyped,
    ExamTypeToJSON,
    ExamTypeToJSONTyped,
} from './ExamType';
import type { SubjectScore } from './SubjectScore';
import {
    SubjectScoreFromJSON,
    SubjectScoreFromJSONTyped,
    SubjectScoreToJSON,
    SubjectScoreToJSONTyped,
} from './SubjectScore';

/**
 * 
 * @export
 * @interface ExamCreate
 */
export interface ExamCreate {
    /**
     * 
     * @type {string}
     * @memberof ExamCreate
     */
    studentId: string;
    /**
     * 
     * @type {string}
     * @memberof ExamCreate
     */
    academicTermId?: string;
    /**
     * 
     * @type {string}
     * @memberof ExamCreate
     */
    examName: string;
    /**
     * 
     * @type {ExamType}
     * @memberof ExamCreate
     */
    examType: ExamType;
    /**
     * 
     * @type {Date}
     * @memberof ExamCreate
     */
    examDate: Date;
    /**
     * 
     * @type {Array<SubjectScore>}
     * @memberof ExamCreate
     */
    subjects: Array<SubjectScore>;
}



/**
 * Check if a given object implements the ExamCreate interface.
 */
export function instanceOfExamCreate(value: object): value is ExamCreate {
    if (!('studentId' in value) || value['studentId'] === undefined) return false;
    if (!('examName' in value) || value['examName'] === undefined) return false;
    if (!('examType' in value) || value['examType'] === undefined) return false;
    if (!('examDate' in value) || value['examDate'] === undefined) return false;
    if (!('subjects' in value) || value['subjects'] === undefined) return false;
    return true;
}

export function ExamCreateFromJSON(json: any): ExamCreate {
    return ExamCreateFromJSONTyped(json, false);
}

export function ExamCreateFromJSONTyped(json: any, ignoreDiscriminator: boolean): ExamCreate {
    if (json == null) {
        return json;
    }
    return {
        
        'studentId': json['studentId'],
        'academicTermId': json['academicTermId'] == null ? undefined : json['academicTermId'],
        'examName': json['examName'],
        'examType': ExamTypeFromJSON(json['examType']),
        'examDate': (new Date(json['examDate'])),
        'subjects': ((json['subjects'] as Array<any>).map(SubjectScoreFromJSON)),
    };
}

export function ExamCreateToJSON(json: any): ExamCreate {
    return ExamCreateToJSONTyped(json, false);
}

export function ExamCreateToJSONTyped(value?: ExamCreate | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'studentId': value['studentId'],
        'academicTermId': value['academicTermId'],
        'examName': value['examName'],
        'examType': ExamTypeToJSON(value['examType']),
        'examDate': ((value['examDate']).toISOString().substring(0,10)),
        'subjects': ((value['subjects'] as Array<any>).map(SubjectScoreToJSON)),
    };
}