// student-growth-system-design-v3/student-growth-system-design-v3/frontend/src/api/generated/models/ExamUpdate.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Student Growth Archive API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 3.15.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { ExamType } from './ExamType';
import {
    ExamTypeFromJSON,
    ExamTypeFromJSONTyped,
    ExamTypeToJSON,
    ExamTypeToJSONTyped,
} from './ExamType';
import type { SubjectScore } from './SubjectScore';
import {
    SubjectScoreFromJSON,
    SubjectScoreFromJSONTyped,
    SubjectScoreToJSON,
    SubjectScoreToJSONTyped,
} from './SubjectScore';

/**
 * 
 * @export
 * @interface ExamUpdate
 */
export interface ExamUpdate {
    /**
     * 
     * @type {string}
     * @memberof ExamUpdate
     */
    studentId: string;
    /**
     * 
     * @type {string}
     * @memberof ExamUpdate
     */
    academicTermId?: string;
    /**
     * 
     * @type {string}
     * @memberof ExamUpdate
     */
    examName: string;
    /**
     * 
     * @type {ExamType}
     * @memberof ExamUpdate
     */
    examType: ExamType;
    /**
     * 
     * @type {Date}
     * @memberof ExamUpdate
     */
    examDate: Date;
    /**
     * 
     * @type {Array<SubjectScore>}
     * @memberof ExamUpdate
     */
    subjects: Array<SubjectScore>;
    /**
     * 
     * @type {number}
     * @memberof ExamUpdate
     */
    version: number;
}



/**
 * Check if a given object implements the ExamUpdate interface.
 */
export function instanceOfExamUpdate(value: object): value is ExamUpdate {
    if (!('studentId' in value) || value['studentId'] === undefined) return false;
    if (!('examName' in value) || value['examName'] === undefined) return false;
    if (!('examType' in value) || value['examType'] === undefined) return false;
    if (!('examDate' in value) || value['examDate'] === undefined) return false;
    if (!('subjects' in value) || value['subjects'] === undefined) return false;
    if (!('version' in value) || value['version'] === undefined) return false;
    return true;
}

export function ExamUpdateFromJSON(json: any): ExamUpdate {
    return ExamUpdateFromJSONTyped(json, false);
}

export function ExamUpdateFromJSONTyped(json: any, ignoreDiscriminator: boolean): ExamUpdate {
    if (json == null) {
        return json;
    }
    return {
        
        'studentId': json['studentId'],
        'academicTermId': json['academicTermId'] == null ? undefined : json['academicTermId'],
        'examName': json['examName'],
        'examType': ExamTypeFromJSON(json['examType']),
        'examDate': (new Date(json['examDate'])),
        'subjects': ((json['subjects'] as Array<any>).map(SubjectScoreFromJSON)),
        'version': json['version'],
    };
}

export function ExamUpdateToJSON(json: any): ExamUpdate {
    return ExamUpdateToJSONTyped(json, false);
}

export function ExamUpdateToJSONTyped(value?: ExamUpdate | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'studentId': value['studentId'],
        'academicTermId': value['academicTermId'],
        'examName': value['examName'],
        'examType': ExamTypeToJSON(value['examType']),
        'examDate': ((value['examDate']).toISOString().substring(0,10)),
        'subjects': ((value['subjects'] as Array<any>).map(SubjectScoreToJSON)),
        'version': value['version'],
    };
}