// student-growth-system-design-v3/student-growth-system-design-v3/frontend/src/api/generated/models/ScoreListItemDto.ts
export interface ScoreListItemDto {
    /**
     * 
     * @type {string}
     * @memberof ScoreListItemDto
     */
    id: string;
    /**
     * 
     * @type {string}
     * @memberof ScoreListItemDto
     */
    examId: string;
    /**
     * 
     * @type {string}
     * @memberof ScoreListItemDto
     */
    examName: string;
    /**
     * 
     * @type {ExamType}
     * @memberof ScoreListItemDto
     */
    examType: ExamType;
    /**
     * 
     * @type {Date}
     * @memberof ScoreListItemDto
     */
    examDate: Date;
    /**
     * 
     * @type {string}
     * @memberof ScoreListItemDto
     */
    academicTermId?: string | null;
    /**
     * 
     * @type {string}
     * @memberof ScoreListItemDto
     */
    subjectId: string;
    /**
     * 
     * @type {string}
     * @memberof ScoreListItemDto
     */
    subjectName: string;
    /**
     * 
     * @type {number}
     * @memberof ScoreListItemDto
     */
    score: number;
    /**
     * 
     * @type {number}
     * @memberof ScoreListItemDto
     */
    fullScore: number;
    /**
     * 
     * @type {number}
     * @memberof ScoreListItemDto
     */
    scoreRate: number;
    /**
     * 
     * @type {number}
     * @memberof ScoreListItemDto
     */
    classRank?: number | null;
    /**
     * 
     * @type {number}
     * @memberof ScoreListItemDto
     */
    gradeRank?: number | null;
}